renamed rfc-cron to cron
It is more consistent
